[svn-r9727] Purpose:     Bug Fix/Code Cleanup/Doc Cleanup/Optimization/Branch Sync :-)
Description:
    Generally speaking, this is the "signed->unsigned" change to selections.
However, in the process of merging code back, things got stickier and stickier
until I ended up doing a big "sync the two branches up" operation.  So... I
brought back all the "infrastructure" fixes from the development branch to the
release branch (which I think were actually making some improvement in
performance) as well as fixed several bugs which had been fixed in one branch,
but not the other.

    I've also tagged the repository before making this checkin with the label
"before_signed_unsigned_changes".

Platforms tested:
    FreeBSD 4.10 (sleipnir) w/parallel & fphdf5
    FreeBSD 4.10 (sleipnir) w/threadsafe
    FreeBSD 4.10 (sleipnir) w/backward compatibility
    Solaris 2.7 (arabica) w/"purify options"
    Solaris 2.8 (sol) w/FORTRAN & C++
    AIX 5.x (copper) w/parallel & FORTRAN
    IRIX64 6.5 (modi4) w/FORTRAN
    Linux 2.4 (heping) w/FORTRAN & C++


Misc. update:

+++ src/H5HGprivate.h
@@ -34,7 +34,7 @@
 /* Information to locate object in global heap */
 typedef struct H5HG_t {
     haddr_t		addr;		/*address of collection		*/
-    unsigned		idx;		/*object ID within collection	*/
+    size_t		idx;		/*object ID within collection	*/
 } H5HG_t;
 
 /* Typedef for heap in memory (defined in H5HGpkg.h) */
